style: exrtract button css to global
